Create Search.js
The Search component will allow users to search for videos by title or description. The component will have a search bar and a button to submit the search. When the user submits the search, the component will call an API endpoint to get the search results and display them on the page.