Create contracts.js
In the "contracts.js" file, we import the contract artifacts for the CreditToken, NFTSubscription, and RewardSystem contracts from their respective JSON files. We then use the Web3 library to connect to the blockchain network and get the deployed contract addresses for the current network. We use these addresses to create new instances of the contracts using the Web3 contract object, which we then return as an object.