Create web3.js
In the "web3.js" file, we first check if the user has MetaMask or another Ethereum-enabled browser extension installed. If so, we use the Web3 provider provided by the extension to connect to the network. If not, we check if there is an injected Web3 instance from the browser. If not, we connect to the network using an HTTP provider at the URL specified in the ".env" file. We then return the Web3 object.